added

+++ base/views.py
@@ -1,11 +1,16 @@
+from email import message
+
 from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.forms import UserCreationForm
 from django.contrib.auth.models import User
 from django.db.models import Q
+from django.http import HttpResponse
 from django.shortcuts import redirect, render
 
 from .forms import RoomForm
-from .models import Room, Topic
+from .models import Message, Room, Topic
 
 # rooms = [
 #     {"id": 1, "name": "lets learn python"},
@@ -15,8 +20,11 @@ from .models import Room, Topic
 
 
 def loginPage(request):
+    page = "login"
+    if request.user.is_authenticated:
+        return redirect("home")
     if request.method == "POST":
-        username = request.POST.get("username")
+        username = request.POST.get("username").lower()
         password = request.POST.get("password")
 
         try:
@@ -31,13 +39,30 @@ def loginPage(request):
         else:
             messages.error(request, "username or password doe not exit")
 
-    context = {}
+    context = {"page": page}
     return render(request, "base/login_register.html", context)
 
 
 def logoutUser(request):
     logout(request)
     return redirect("home")
+
+
+def registerPage(request):
+    form = UserCreationForm()
+
+    if request.method == "POST":
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect("home")
+        else:
+            messages.error(request, "An error Occured during registeration")
+
+    return render(request, "base/login_register.html", {"form": form})
 
 
 def home(request):
@@ -53,13 +78,27 @@ def home(request):
 
 def room(request, pk):
     room = Room.objects.get(id=pk)
-    context = {"room": room}
+    room_messages = room.message_set.all().order_by("-created")
+    participants = room.participants.all()
+
+    if request.method == "POST":
+        message = Message.objects.create(
+            user=request.user, room=room, body=request.POST.get("body")
+        )
+        room.participants.add(request.user)
+        return redirect("room", pk=room.id)
+
+    context = {
+        "room": room,
+        "room_messages": room_messages,
+        "participants": participants,
+    }
     return render(request, "base/room.html", context)
 
 
+@login_required(login_url="login")
 def createRoom(request):
     form = RoomForm()
-
     if request.method == "POST":
         form = RoomForm(request.POST)
         if form.is_valid():
@@ -69,9 +108,13 @@ def createRoom(request):
     return render(request, "base/room_form.html", context)
 
 
+@login_required(login_url="login")
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("you are not allowed here!!!")
 
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room)
@@ -83,9 +126,25 @@ def updateRoom(request, pk):
     return render(request, "base/room_form.html", context)
 
 
+@login_required(login_url="login")
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse("you are not allowed here!!!")
+
     if request.method == "POST":
         room.delete()
         return redirect("home")
     return render(request, "base/delete.html", {"obj": room})
+
+
+@login_required(login_url="login")
+def deleteMessage(request, pk):
+    message = Message.objects.get(id=pk)
+    if request.user != message.user:
+        return HttpResponse("you are not allowed here!!!")
+
+    if request.method == "POST":
+        message.delete()
+        return redirect("home")
+    return render(request, "base/delete.html", {"obj": message})
